+++ toolsrc/include/vcpkg/commands.hash.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <vcpkg/commands.interface.h>
+
+namespace vcpkg::Commands::Hash
+{
+    void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
+}
