+++ toolsrc/include/vcpkg/commands.interface.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <vcpkg/vcpkgcmdarguments.h>
+#include <vcpkg/vcpkgpaths.h>
+
+namespace vcpkg::Commands
+{
+    enum class DryRun : bool
+    {
+        No,
+        Yes,
+    };
+}
