+++ toolsrc/include/vcpkg/commands.list.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <vcpkg/commands.interface.h>
+
+namespace vcpkg::Commands::List
+{
+    extern const CommandStructure COMMAND_STRUCTURE;
+    void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
+}
