+++ toolsrc/src/vcpkg/commands.buildexternal.cpp
@@ -3,7 +3,7 @@
 #include <vcpkg/binarycaching.h>
 #include <vcpkg/build.h>
 #include <vcpkg/cmakevars.h>
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.buildexternal.h>
 #include <vcpkg/help.h>
 #include <vcpkg/input.h>
