+++ toolsrc/src/vcpkg/commands.contact.cpp
@@ -4,7 +4,7 @@
 #include <vcpkg/base/system.print.h>
 #include <vcpkg/base/system.process.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.contact.h>
 #include <vcpkg/help.h>
 #include <vcpkg/userconfig.h>
