+++ toolsrc/src/vcpkg/commands.create.cpp
@@ -4,7 +4,7 @@
 #include <vcpkg/base/files.h>
 
 #include <vcpkg/buildenvironment.h>
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.create.h>
 #include <vcpkg/help.h>
 
 namespace vcpkg::Commands::Create
