+++ toolsrc/src/vcpkg/commands.dependinfo.cpp
@@ -4,7 +4,7 @@
 #include <vcpkg/base/system.print.h>
 #include <vcpkg/base/util.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.dependinfo.h>
 #include <vcpkg/dependencies.h>
 #include <vcpkg/help.h>
 #include <vcpkg/input.h>
