+++ toolsrc/src/vcpkg/commands.integrate.cpp
@@ -7,7 +7,7 @@
 #include <vcpkg/base/system.process.h>
 #include <vcpkg/base/util.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.integrate.h>
 #include <vcpkg/metrics.h>
 #include <vcpkg/userconfig.h>
