+++ toolsrc/src/vcpkg/commands.list.cpp
@@ -2,7 +2,7 @@
 
 #include <vcpkg/base/system.print.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.list.h>
 #include <vcpkg/help.h>
 #include <vcpkg/vcpkglib.h>
 #include <vcpkg/versiont.h>
