+++ toolsrc/src/vcpkg/commands.owns.cpp
@@ -2,7 +2,7 @@
 
 #include <vcpkg/base/system.print.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.owns.h>
 #include <vcpkg/help.h>
 #include <vcpkg/vcpkglib.h>
