+++ toolsrc/src/vcpkg/commands.porthistory.cpp
@@ -4,7 +4,7 @@
 #include <vcpkg/base/system.process.h>
 #include <vcpkg/base/util.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.porthistory.h>
 #include <vcpkg/help.h>
 
 namespace vcpkg::Commands::PortHistory
