+++ toolsrc/src/vcpkg/commands.portsdiff.cpp
@@ -5,7 +5,7 @@
 #include <vcpkg/base/system.process.h>
 #include <vcpkg/base/util.h>
 
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.portsdiff.h>
 #include <vcpkg/help.h>
 #include <vcpkg/paragraphs.h>
 #include <vcpkg/versiont.h>
