+++ toolsrc/src/vcpkg/commands.setinstalled.cpp
@@ -3,7 +3,7 @@
 #include <vcpkg/base/system.print.h>
 
 #include <vcpkg/binarycaching.h>
-#include <vcpkg/commands.h>
+#include <vcpkg/commands.setinstalled.h>
 #include <vcpkg/globalstate.h>
 #include <vcpkg/help.h>
 #include <vcpkg/input.h>
