+++ toolsrc/src/vcpkg/vcpkgcmdarguments.cpp
@@ -4,6 +4,7 @@
 #include <vcpkg/base/system.print.h>
 
 #include <vcpkg/commands.h>
+#include <vcpkg/commands.integrate.h>
 #include <vcpkg/globalstate.h>
 #include <vcpkg/metrics.h>
 #include <vcpkg/vcpkgcmdarguments.h>
