+++ toolsrc/vcpkglib/vcpkglib.vcxproj
@@ -180,6 +180,30 @@
     <ClInclude Include="..\include\vcpkg\buildenvironment.h" />
     <ClInclude Include="..\include\vcpkg\cmakevars.h" />
     <ClInclude Include="..\include\vcpkg\commands.h" />
+    <ClInclude Include="..\include\vcpkg\commands.autocomplete.h" />
+    <ClInclude Include="..\include\vcpkg\commands.buildexternal.h" />
+    <ClInclude Include="..\include\vcpkg\commands.cache.h" />
+    <ClInclude Include="..\include\vcpkg\commands.ci.h" />
+    <ClInclude Include="..\include\vcpkg\commands.ciclean.h" />
+    <ClInclude Include="..\include\vcpkg\commands.contact.h" />
+    <ClInclude Include="..\include\vcpkg\commands.create.h" />
+    <ClInclude Include="..\include\vcpkg\commands.dependinfo.h" />
+    <ClInclude Include="..\include\vcpkg\commands.edit.h" />
+    <ClInclude Include="..\include\vcpkg\commands.env.h" />
+    <ClInclude Include="..\include\vcpkg\commands.fetch.h" />
+    <ClInclude Include="..\include\vcpkg\commands.format-manifest.h" />
+    <ClInclude Include="..\include\vcpkg\commands.hash.h" />
+    <ClInclude Include="..\include\vcpkg\commands.integrate.h" />
+    <ClInclude Include="..\include\vcpkg\commands.interface.h" />
+    <ClInclude Include="..\include\vcpkg\commands.list.h" />
+    <ClInclude Include="..\include\vcpkg\commands.owns.h" />
+    <ClInclude Include="..\include\vcpkg\commands.porthistory.h" />
+    <ClInclude Include="..\include\vcpkg\commands.portsdiff.h" />
+    <ClInclude Include="..\include\vcpkg\commands.search.h" />
+    <ClInclude Include="..\include\vcpkg\commands.setinstalled.h" />
+    <ClInclude Include="..\include\vcpkg\commands.upgrade.h" />
+    <ClInclude Include="..\include\vcpkg\commands.version.h" />
+    <ClInclude Include="..\include\vcpkg\commands.xvsinstances.h" />
     <ClInclude Include="..\include\vcpkg\dependencies.h" />
     <ClInclude Include="..\include\vcpkg\export.chocolatey.h" />
     <ClInclude Include="..\include\vcpkg\export.h" />
@@ -253,8 +277,9 @@
     <ClCompile Include="..\src\vcpkg\commands.dependinfo.cpp" />
     <ClCompile Include="..\src\vcpkg\commands.edit.cpp" />
     <ClCompile Include="..\src\vcpkg\commands.env.cpp" />
-    <ClCompile Include="..\src\vcpkg\commands.exportifw.cpp" />
+    <ClCompile Include="..\src\vcpkg\commands.fetch.cpp" />
     <ClCompile Include="..\src\vcpkg\commands.format-manifest.cpp" />
+    <ClCompile Include="..\src\vcpkg\commands.hash.cpp" />
     <ClCompile Include="..\src\vcpkg\commands.integrate.cpp" />
     <ClCompile Include="..\src\vcpkg\commands.list.cpp" />
     <ClCompile Include="..\src\vcpkg\commands.owns.cpp" />
@@ -268,6 +293,7 @@
     <ClCompile Include="..\src\vcpkg\dependencies.cpp" />
     <ClCompile Include="..\src\vcpkg\export.cpp" />
     <ClCompile Include="..\src\vcpkg\export.chocolatey.cpp" />
+    <ClCompile Include="..\src\vcpkg\export.ifw.cpp" />
     <ClCompile Include="..\src\vcpkg\export.prefab.cpp" />
     <ClCompile Include="..\src\vcpkg\globalstate.cpp" />
     <ClCompile Include="..\src\vcpkg\help.cpp" />
